Relational data records

Generic function, for retrieving data contained in a database-like structure. In particular, this is intended for such structures where the individual relations can't be accessed with subsetting.

Usage

records(x, ...)

records(x, ...) <- value

Arguments

x

a relational data object, such as a relation or database object.

...

further arguments passed on to methods.

value

A list of data frames of the same length as records(x, ...), where each data frame has the same column names as that which it will replace, in the same order.

Value

A list containing data frames, with elements named for their respective relations.

Details

Since the relational data objects in autodb, relation and database, have subsetting methods that return relational data objects, the data contained within them can't be accessed by subsetting. This function is intended for accessing it instead.

It's recommended to call records before doing any subsetting, since subsetting on a relation data object does more work that will be thrown away, such as subsetting on a database checking whether foreign key references should be removed.

Examples

db <- autodb(ChickWeight)
records(db) # data for Chick and Time_Chick relations
#> $Chick
#>     Chick Diet
#> 1       1    1
#> 13      2    1
#> 25      3    1
#> 37      4    1
#> 49      5    1
#> 61      6    1
#> 73      7    1
#> 85      8    1
#> 96      9    1
#> 108    10    1
#> 120    11    1
#> 132    12    1
#> 144    13    1
#> 156    14    1
#> 168    15    1
#> 176    16    1
#> 183    17    1
#> 195    18    1
#> 197    19    1
#> 209    20    1
#> 221    21    2
#> 233    22    2
#> 245    23    2
#> 257    24    2
#> 269    25    2
#> 281    26    2
#> 293    27    2
#> 305    28    2
#> 317    29    2
#> 329    30    2
#> 341    31    3
#> 353    32    3
#> 365    33    3
#> 377    34    3
#> 389    35    3
#> 401    36    3
#> 413    37    3
#> 425    38    3
#> 437    39    3
#> 449    40    3
#> 461    41    4
#> 473    42    4
#> 485    43    4
#> 497    44    4
#> 507    45    4
#> 519    46    4
#> 531    47    4
#> 543    48    4
#> 555    49    4
#> 567    50    4
#> 
#> $Time_Chick
#>     Time Chick weight
#> 1      0     1     42
#> 2      2     1     51
#> 3      4     1     59
#> 4      6     1     64
#> 5      8     1     76
#> 6     10     1     93
#> 7     12     1    106
#> 8     14     1    125
#> 9     16     1    149
#> 10    18     1    171
#> 11    20     1    199
#> 12    21     1    205
#> 13     0     2     40
#> 14     2     2     49
#> 15     4     2     58
#> 16     6     2     72
#> 17     8     2     84
#> 18    10     2    103
#> 19    12     2    122
#> 20    14     2    138
#> 21    16     2    162
#> 22    18     2    187
#> 23    20     2    209
#> 24    21     2    215
#> 25     0     3     43
#> 26     2     3     39
#> 27     4     3     55
#> 28     6     3     67
#> 29     8     3     84
#> 30    10     3     99
#> 31    12     3    115
#> 32    14     3    138
#> 33    16     3    163
#> 34    18     3    187
#> 35    20     3    198
#> 36    21     3    202
#> 37     0     4     42
#> 38     2     4     49
#> 39     4     4     56
#> 40     6     4     67
#> 41     8     4     74
#> 42    10     4     87
#> 43    12     4    102
#> 44    14     4    108
#> 45    16     4    136
#> 46    18     4    154
#> 47    20     4    160
#> 48    21     4    157
#> 49     0     5     41
#> 50     2     5     42
#> 51     4     5     48
#> 52     6     5     60
#> 53     8     5     79
#> 54    10     5    106
#> 55    12     5    141
#> 56    14     5    164
#> 57    16     5    197
#> 58    18     5    199
#> 59    20     5    220
#> 60    21     5    223
#> 61     0     6     41
#> 62     2     6     49
#> 63     4     6     59
#> 64     6     6     74
#> 65     8     6     97
#> 66    10     6    124
#> 67    12     6    141
#> 68    14     6    148
#> 69    16     6    155
#> 70    18     6    160
#> 71    20     6    160
#> 72    21     6    157
#> 73     0     7     41
#> 74     2     7     49
#> 75     4     7     57
#> 76     6     7     71
#> 77     8     7     89
#> 78    10     7    112
#> 79    12     7    146
#> 80    14     7    174
#> 81    16     7    218
#> 82    18     7    250
#> 83    20     7    288
#> 84    21     7    305
#> 85     0     8     42
#> 86     2     8     50
#> 87     4     8     61
#> 88     6     8     71
#> 89     8     8     84
#> 90    10     8     93
#> 91    12     8    110
#> 92    14     8    116
#> 93    16     8    126
#> 94    18     8    134
#> 95    20     8    125
#> 96     0     9     42
#> 97     2     9     51
#> 98     4     9     59
#> 99     6     9     68
#> 100    8     9     85
#> 101   10     9     96
#> 102   12     9     90
#> 103   14     9     92
#> 104   16     9     93
#> 105   18     9    100
#> 106   20     9    100
#> 107   21     9     98
#> 108    0    10     41
#> 109    2    10     44
#> 110    4    10     52
#> 111    6    10     63
#> 112    8    10     74
#> 113   10    10     81
#> 114   12    10     89
#> 115   14    10     96
#> 116   16    10    101
#> 117   18    10    112
#> 118   20    10    120
#> 119   21    10    124
#> 120    0    11     43
#> 121    2    11     51
#> 122    4    11     63
#> 123    6    11     84
#> 124    8    11    112
#> 125   10    11    139
#> 126   12    11    168
#> 127   14    11    177
#> 128   16    11    182
#> 129   18    11    184
#> 130   20    11    181
#> 131   21    11    175
#> 132    0    12     41
#> 133    2    12     49
#> 134    4    12     56
#> 135    6    12     62
#> 136    8    12     72
#> 137   10    12     88
#> 138   12    12    119
#> 139   14    12    135
#> 140   16    12    162
#> 141   18    12    185
#> 142   20    12    195
#> 143   21    12    205
#> 144    0    13     41
#> 145    2    13     48
#> 146    4    13     53
#> 147    6    13     60
#> 148    8    13     65
#> 149   10    13     67
#> 150   12    13     71
#> 151   14    13     70
#> 152   16    13     71
#> 153   18    13     81
#> 154   20    13     91
#> 155   21    13     96
#> 156    0    14     41
#> 157    2    14     49
#> 158    4    14     62
#> 159    6    14     79
#> 160    8    14    101
#> 161   10    14    128
#> 162   12    14    164
#> 163   14    14    192
#> 164   16    14    227
#> 165   18    14    248
#> 166   20    14    259
#> 167   21    14    266
#> 168    0    15     41
#> 169    2    15     49
#> 170    4    15     56
#> 171    6    15     64
#> 172    8    15     68
#> 173   10    15     68
#> 174   12    15     67
#> 175   14    15     68
#> 176    0    16     41
#> 177    2    16     45
#> 178    4    16     49
#> 179    6    16     51
#> 180    8    16     57
#> 181   10    16     51
#> 182   12    16     54
#> 183    0    17     42
#> 184    2    17     51
#> 185    4    17     61
#> 186    6    17     72
#> 187    8    17     83
#> 188   10    17     89
#> 189   12    17     98
#> 190   14    17    103
#> 191   16    17    113
#> 192   18    17    123
#> 193   20    17    133
#> 194   21    17    142
#> 195    0    18     39
#> 196    2    18     35
#> 197    0    19     43
#> 198    2    19     48
#> 199    4    19     55
#> 200    6    19     62
#> 201    8    19     65
#> 202   10    19     71
#> 203   12    19     82
#> 204   14    19     88
#> 205   16    19    106
#> 206   18    19    120
#> 207   20    19    144
#> 208   21    19    157
#> 209    0    20     41
#> 210    2    20     47
#> 211    4    20     54
#> 212    6    20     58
#> 213    8    20     65
#> 214   10    20     73
#> 215   12    20     77
#> 216   14    20     89
#> 217   16    20     98
#> 218   18    20    107
#> 219   20    20    115
#> 220   21    20    117
#> 221    0    21     40
#> 222    2    21     50
#> 223    4    21     62
#> 224    6    21     86
#> 225    8    21    125
#> 226   10    21    163
#> 227   12    21    217
#> 228   14    21    240
#> 229   16    21    275
#> 230   18    21    307
#> 231   20    21    318
#> 232   21    21    331
#> 233    0    22     41
#> 234    2    22     55
#> 235    4    22     64
#> 236    6    22     77
#> 237    8    22     90
#> 238   10    22     95
#> 239   12    22    108
#> 240   14    22    111
#> 241   16    22    131
#> 242   18    22    148
#> 243   20    22    164
#> 244   21    22    167
#> 245    0    23     43
#> 246    2    23     52
#> 247    4    23     61
#> 248    6    23     73
#> 249    8    23     90
#> 250   10    23    103
#> 251   12    23    127
#> 252   14    23    135
#> 253   16    23    145
#> 254   18    23    163
#> 255   20    23    170
#> 256   21    23    175
#> 257    0    24     42
#> 258    2    24     52
#> 259    4    24     58
#> 260    6    24     74
#> 261    8    24     66
#> 262   10    24     68
#> 263   12    24     70
#> 264   14    24     71
#> 265   16    24     72
#> 266   18    24     72
#> 267   20    24     76
#> 268   21    24     74
#> 269    0    25     40
#> 270    2    25     49
#> 271    4    25     62
#> 272    6    25     78
#> 273    8    25    102
#> 274   10    25    124
#> 275   12    25    146
#> 276   14    25    164
#> 277   16    25    197
#> 278   18    25    231
#> 279   20    25    259
#> 280   21    25    265
#> 281    0    26     42
#> 282    2    26     48
#> 283    4    26     57
#> 284    6    26     74
#> 285    8    26     93
#> 286   10    26    114
#> 287   12    26    136
#> 288   14    26    147
#> 289   16    26    169
#> 290   18    26    205
#> 291   20    26    236
#> 292   21    26    251
#> 293    0    27     39
#> 294    2    27     46
#> 295    4    27     58
#> 296    6    27     73
#> 297    8    27     87
#> 298   10    27    100
#> 299   12    27    115
#> 300   14    27    123
#> 301   16    27    144
#> 302   18    27    163
#> 303   20    27    185
#> 304   21    27    192
#> 305    0    28     39
#> 306    2    28     46
#> 307    4    28     58
#> 308    6    28     73
#> 309    8    28     92
#> 310   10    28    114
#> 311   12    28    145
#> 312   14    28    156
#> 313   16    28    184
#> 314   18    28    207
#> 315   20    28    212
#> 316   21    28    233
#> 317    0    29     39
#> 318    2    29     48
#> 319    4    29     59
#> 320    6    29     74
#> 321    8    29     87
#> 322   10    29    106
#> 323   12    29    134
#> 324   14    29    150
#> 325   16    29    187
#> 326   18    29    230
#> 327   20    29    279
#> 328   21    29    309
#> 329    0    30     42
#> 330    2    30     48
#> 331    4    30     59
#> 332    6    30     72
#> 333    8    30     85
#> 334   10    30     98
#> 335   12    30    115
#> 336   14    30    122
#> 337   16    30    143
#> 338   18    30    151
#> 339   20    30    157
#> 340   21    30    150
#> 341    0    31     42
#> 342    2    31     53
#> 343    4    31     62
#> 344    6    31     73
#> 345    8    31     85
#> 346   10    31    102
#> 347   12    31    123
#> 348   14    31    138
#> 349   16    31    170
#> 350   18    31    204
#> 351   20    31    235
#> 352   21    31    256
#> 353    0    32     41
#> 354    2    32     49
#> 355    4    32     65
#> 356    6    32     82
#> 357    8    32    107
#> 358   10    32    129
#> 359   12    32    159
#> 360   14    32    179
#> 361   16    32    221
#> 362   18    32    263
#> 363   20    32    291
#> 364   21    32    305
#> 365    0    33     39
#> 366    2    33     50
#> 367    4    33     63
#> 368    6    33     77
#> 369    8    33     96
#> 370   10    33    111
#> 371   12    33    137
#> 372   14    33    144
#> 373   16    33    151
#> 374   18    33    146
#> 375   20    33    156
#> 376   21    33    147
#> 377    0    34     41
#> 378    2    34     49
#> 379    4    34     63
#> 380    6    34     85
#> 381    8    34    107
#> 382   10    34    134
#> 383   12    34    164
#> 384   14    34    186
#> 385   16    34    235
#> 386   18    34    294
#> 387   20    34    327
#> 388   21    34    341
#> 389    0    35     41
#> 390    2    35     53
#> 391    4    35     64
#> 392    6    35     87
#> 393    8    35    123
#> 394   10    35    158
#> 395   12    35    201
#> 396   14    35    238
#> 397   16    35    287
#> 398   18    35    332
#> 399   20    35    361
#> 400   21    35    373
#> 401    0    36     39
#> 402    2    36     48
#> 403    4    36     61
#> 404    6    36     76
#> 405    8    36     98
#> 406   10    36    116
#> 407   12    36    145
#> 408   14    36    166
#> 409   16    36    198
#> 410   18    36    227
#> 411   20    36    225
#> 412   21    36    220
#> 413    0    37     41
#> 414    2    37     48
#> 415    4    37     56
#> 416    6    37     68
#> 417    8    37     80
#> 418   10    37     83
#> 419   12    37    103
#> 420   14    37    112
#> 421   16    37    135
#> 422   18    37    157
#> 423   20    37    169
#> 424   21    37    178
#> 425    0    38     41
#> 426    2    38     49
#> 427    4    38     61
#> 428    6    38     74
#> 429    8    38     98
#> 430   10    38    109
#> 431   12    38    128
#> 432   14    38    154
#> 433   16    38    192
#> 434   18    38    232
#> 435   20    38    280
#> 436   21    38    290
#> 437    0    39     42
#> 438    2    39     50
#> 439    4    39     61
#> 440    6    39     78
#> 441    8    39     89
#> 442   10    39    109
#> 443   12    39    130
#> 444   14    39    146
#> 445   16    39    170
#> 446   18    39    214
#> 447   20    39    250
#> 448   21    39    272
#> 449    0    40     41
#> 450    2    40     55
#> 451    4    40     66
#> 452    6    40     79
#> 453    8    40    101
#> 454   10    40    120
#> 455   12    40    154
#> 456   14    40    182
#> 457   16    40    215
#> 458   18    40    262
#> 459   20    40    295
#> 460   21    40    321
#> 461    0    41     42
#> 462    2    41     51
#> 463    4    41     66
#> 464    6    41     85
#> 465    8    41    103
#> 466   10    41    124
#> 467   12    41    155
#> 468   14    41    153
#> 469   16    41    175
#> 470   18    41    184
#> 471   20    41    199
#> 472   21    41    204
#> 473    0    42     42
#> 474    2    42     49
#> 475    4    42     63
#> 476    6    42     84
#> 477    8    42    103
#> 478   10    42    126
#> 479   12    42    160
#> 480   14    42    174
#> 481   16    42    204
#> 482   18    42    234
#> 483   20    42    269
#> 484   21    42    281
#> 485    0    43     42
#> 486    2    43     55
#> 487    4    43     69
#> 488    6    43     96
#> 489    8    43    131
#> 490   10    43    157
#> 491   12    43    184
#> 492   14    43    188
#> 493   16    43    197
#> 494   18    43    198
#> 495   20    43    199
#> 496   21    43    200
#> 497    0    44     42
#> 498    2    44     51
#> 499    4    44     65
#> 500    6    44     86
#> 501    8    44    103
#> 502   10    44    118
#> 503   12    44    127
#> 504   14    44    138
#> 505   16    44    145
#> 506   18    44    146
#> 507    0    45     41
#> 508    2    45     50
#> 509    4    45     61
#> 510    6    45     78
#> 511    8    45     98
#> 512   10    45    117
#> 513   12    45    135
#> 514   14    45    141
#> 515   16    45    147
#> 516   18    45    174
#> 517   20    45    197
#> 518   21    45    196
#> 519    0    46     40
#> 520    2    46     52
#> 521    4    46     62
#> 522    6    46     82
#> 523    8    46    101
#> 524   10    46    120
#> 525   12    46    144
#> 526   14    46    156
#> 527   16    46    173
#> 528   18    46    210
#> 529   20    46    231
#> 530   21    46    238
#> 531    0    47     41
#> 532    2    47     53
#> 533    4    47     66
#> 534    6    47     79
#> 535    8    47    100
#> 536   10    47    123
#> 537   12    47    148
#> 538   14    47    157
#> 539   16    47    168
#> 540   18    47    185
#> 541   20    47    210
#> 542   21    47    205
#> 543    0    48     39
#> 544    2    48     50
#> 545    4    48     62
#> 546    6    48     80
#> 547    8    48    104
#> 548   10    48    125
#> 549   12    48    154
#> 550   14    48    170
#> 551   16    48    222
#> 552   18    48    261
#> 553   20    48    303
#> 554   21    48    322
#> 555    0    49     40
#> 556    2    49     53
#> 557    4    49     64
#> 558    6    49     85
#> 559    8    49    108
#> 560   10    49    128
#> 561   12    49    152
#> 562   14    49    166
#> 563   16    49    184
#> 564   18    49    203
#> 565   20    49    233
#> 566   21    49    237
#> 567    0    50     41
#> 568    2    50     54
#> 569    4    50     67
#> 570    6    50     84
#> 571    8    50    105
#> 572   10    50    122
#> 573   12    50    155
#> 574   14    50    175
#> 575   16    50    205
#> 576   18    50    234
#> 577   20    50    264
#> 578   21    50    264
#> 

# ways to get data for subsets
records(db)[c(1, 2)]
#> $Chick
#>     Chick Diet
#> 1       1    1
#> 13      2    1
#> 25      3    1
#> 37      4    1
#> 49      5    1
#> 61      6    1
#> 73      7    1
#> 85      8    1
#> 96      9    1
#> 108    10    1
#> 120    11    1
#> 132    12    1
#> 144    13    1
#> 156    14    1
#> 168    15    1
#> 176    16    1
#> 183    17    1
#> 195    18    1
#> 197    19    1
#> 209    20    1
#> 221    21    2
#> 233    22    2
#> 245    23    2
#> 257    24    2
#> 269    25    2
#> 281    26    2
#> 293    27    2
#> 305    28    2
#> 317    29    2
#> 329    30    2
#> 341    31    3
#> 353    32    3
#> 365    33    3
#> 377    34    3
#> 389    35    3
#> 401    36    3
#> 413    37    3
#> 425    38    3
#> 437    39    3
#> 449    40    3
#> 461    41    4
#> 473    42    4
#> 485    43    4
#> 497    44    4
#> 507    45    4
#> 519    46    4
#> 531    47    4
#> 543    48    4
#> 555    49    4
#> 567    50    4
#> 
#> $Time_Chick
#>     Time Chick weight
#> 1      0     1     42
#> 2      2     1     51
#> 3      4     1     59
#> 4      6     1     64
#> 5      8     1     76
#> 6     10     1     93
#> 7     12     1    106
#> 8     14     1    125
#> 9     16     1    149
#> 10    18     1    171
#> 11    20     1    199
#> 12    21     1    205
#> 13     0     2     40
#> 14     2     2     49
#> 15     4     2     58
#> 16     6     2     72
#> 17     8     2     84
#> 18    10     2    103
#> 19    12     2    122
#> 20    14     2    138
#> 21    16     2    162
#> 22    18     2    187
#> 23    20     2    209
#> 24    21     2    215
#> 25     0     3     43
#> 26     2     3     39
#> 27     4     3     55
#> 28     6     3     67
#> 29     8     3     84
#> 30    10     3     99
#> 31    12     3    115
#> 32    14     3    138
#> 33    16     3    163
#> 34    18     3    187
#> 35    20     3    198
#> 36    21     3    202
#> 37     0     4     42
#> 38     2     4     49
#> 39     4     4     56
#> 40     6     4     67
#> 41     8     4     74
#> 42    10     4     87
#> 43    12     4    102
#> 44    14     4    108
#> 45    16     4    136
#> 46    18     4    154
#> 47    20     4    160
#> 48    21     4    157
#> 49     0     5     41
#> 50     2     5     42
#> 51     4     5     48
#> 52     6     5     60
#> 53     8     5     79
#> 54    10     5    106
#> 55    12     5    141
#> 56    14     5    164
#> 57    16     5    197
#> 58    18     5    199
#> 59    20     5    220
#> 60    21     5    223
#> 61     0     6     41
#> 62     2     6     49
#> 63     4     6     59
#> 64     6     6     74
#> 65     8     6     97
#> 66    10     6    124
#> 67    12     6    141
#> 68    14     6    148
#> 69    16     6    155
#> 70    18     6    160
#> 71    20     6    160
#> 72    21     6    157
#> 73     0     7     41
#> 74     2     7     49
#> 75     4     7     57
#> 76     6     7     71
#> 77     8     7     89
#> 78    10     7    112
#> 79    12     7    146
#> 80    14     7    174
#> 81    16     7    218
#> 82    18     7    250
#> 83    20     7    288
#> 84    21     7    305
#> 85     0     8     42
#> 86     2     8     50
#> 87     4     8     61
#> 88     6     8     71
#> 89     8     8     84
#> 90    10     8     93
#> 91    12     8    110
#> 92    14     8    116
#> 93    16     8    126
#> 94    18     8    134
#> 95    20     8    125
#> 96     0     9     42
#> 97     2     9     51
#> 98     4     9     59
#> 99     6     9     68
#> 100    8     9     85
#> 101   10     9     96
#> 102   12     9     90
#> 103   14     9     92
#> 104   16     9     93
#> 105   18     9    100
#> 106   20     9    100
#> 107   21     9     98
#> 108    0    10     41
#> 109    2    10     44
#> 110    4    10     52
#> 111    6    10     63
#> 112    8    10     74
#> 113   10    10     81
#> 114   12    10     89
#> 115   14    10     96
#> 116   16    10    101
#> 117   18    10    112
#> 118   20    10    120
#> 119   21    10    124
#> 120    0    11     43
#> 121    2    11     51
#> 122    4    11     63
#> 123    6    11     84
#> 124    8    11    112
#> 125   10    11    139
#> 126   12    11    168
#> 127   14    11    177
#> 128   16    11    182
#> 129   18    11    184
#> 130   20    11    181
#> 131   21    11    175
#> 132    0    12     41
#> 133    2    12     49
#> 134    4    12     56
#> 135    6    12     62
#> 136    8    12     72
#> 137   10    12     88
#> 138   12    12    119
#> 139   14    12    135
#> 140   16    12    162
#> 141   18    12    185
#> 142   20    12    195
#> 143   21    12    205
#> 144    0    13     41
#> 145    2    13     48
#> 146    4    13     53
#> 147    6    13     60
#> 148    8    13     65
#> 149   10    13     67
#> 150   12    13     71
#> 151   14    13     70
#> 152   16    13     71
#> 153   18    13     81
#> 154   20    13     91
#> 155   21    13     96
#> 156    0    14     41
#> 157    2    14     49
#> 158    4    14     62
#> 159    6    14     79
#> 160    8    14    101
#> 161   10    14    128
#> 162   12    14    164
#> 163   14    14    192
#> 164   16    14    227
#> 165   18    14    248
#> 166   20    14    259
#> 167   21    14    266
#> 168    0    15     41
#> 169    2    15     49
#> 170    4    15     56
#> 171    6    15     64
#> 172    8    15     68
#> 173   10    15     68
#> 174   12    15     67
#> 175   14    15     68
#> 176    0    16     41
#> 177    2    16     45
#> 178    4    16     49
#> 179    6    16     51
#> 180    8    16     57
#> 181   10    16     51
#> 182   12    16     54
#> 183    0    17     42
#> 184    2    17     51
#> 185    4    17     61
#> 186    6    17     72
#> 187    8    17     83
#> 188   10    17     89
#> 189   12    17     98
#> 190   14    17    103
#> 191   16    17    113
#> 192   18    17    123
#> 193   20    17    133
#> 194   21    17    142
#> 195    0    18     39
#> 196    2    18     35
#> 197    0    19     43
#> 198    2    19     48
#> 199    4    19     55
#> 200    6    19     62
#> 201    8    19     65
#> 202   10    19     71
#> 203   12    19     82
#> 204   14    19     88
#> 205   16    19    106
#> 206   18    19    120
#> 207   20    19    144
#> 208   21    19    157
#> 209    0    20     41
#> 210    2    20     47
#> 211    4    20     54
#> 212    6    20     58
#> 213    8    20     65
#> 214   10    20     73
#> 215   12    20     77
#> 216   14    20     89
#> 217   16    20     98
#> 218   18    20    107
#> 219   20    20    115
#> 220   21    20    117
#> 221    0    21     40
#> 222    2    21     50
#> 223    4    21     62
#> 224    6    21     86
#> 225    8    21    125
#> 226   10    21    163
#> 227   12    21    217
#> 228   14    21    240
#> 229   16    21    275
#> 230   18    21    307
#> 231   20    21    318
#> 232   21    21    331
#> 233    0    22     41
#> 234    2    22     55
#> 235    4    22     64
#> 236    6    22     77
#> 237    8    22     90
#> 238   10    22     95
#> 239   12    22    108
#> 240   14    22    111
#> 241   16    22    131
#> 242   18    22    148
#> 243   20    22    164
#> 244   21    22    167
#> 245    0    23     43
#> 246    2    23     52
#> 247    4    23     61
#> 248    6    23     73
#> 249    8    23     90
#> 250   10    23    103
#> 251   12    23    127
#> 252   14    23    135
#> 253   16    23    145
#> 254   18    23    163
#> 255   20    23    170
#> 256   21    23    175
#> 257    0    24     42
#> 258    2    24     52
#> 259    4    24     58
#> 260    6    24     74
#> 261    8    24     66
#> 262   10    24     68
#> 263   12    24     70
#> 264   14    24     71
#> 265   16    24     72
#> 266   18    24     72
#> 267   20    24     76
#> 268   21    24     74
#> 269    0    25     40
#> 270    2    25     49
#> 271    4    25     62
#> 272    6    25     78
#> 273    8    25    102
#> 274   10    25    124
#> 275   12    25    146
#> 276   14    25    164
#> 277   16    25    197
#> 278   18    25    231
#> 279   20    25    259
#> 280   21    25    265
#> 281    0    26     42
#> 282    2    26     48
#> 283    4    26     57
#> 284    6    26     74
#> 285    8    26     93
#> 286   10    26    114
#> 287   12    26    136
#> 288   14    26    147
#> 289   16    26    169
#> 290   18    26    205
#> 291   20    26    236
#> 292   21    26    251
#> 293    0    27     39
#> 294    2    27     46
#> 295    4    27     58
#> 296    6    27     73
#> 297    8    27     87
#> 298   10    27    100
#> 299   12    27    115
#> 300   14    27    123
#> 301   16    27    144
#> 302   18    27    163
#> 303   20    27    185
#> 304   21    27    192
#> 305    0    28     39
#> 306    2    28     46
#> 307    4    28     58
#> 308    6    28     73
#> 309    8    28     92
#> 310   10    28    114
#> 311   12    28    145
#> 312   14    28    156
#> 313   16    28    184
#> 314   18    28    207
#> 315   20    28    212
#> 316   21    28    233
#> 317    0    29     39
#> 318    2    29     48
#> 319    4    29     59
#> 320    6    29     74
#> 321    8    29     87
#> 322   10    29    106
#> 323   12    29    134
#> 324   14    29    150
#> 325   16    29    187
#> 326   18    29    230
#> 327   20    29    279
#> 328   21    29    309
#> 329    0    30     42
#> 330    2    30     48
#> 331    4    30     59
#> 332    6    30     72
#> 333    8    30     85
#> 334   10    30     98
#> 335   12    30    115
#> 336   14    30    122
#> 337   16    30    143
#> 338   18    30    151
#> 339   20    30    157
#> 340   21    30    150
#> 341    0    31     42
#> 342    2    31     53
#> 343    4    31     62
#> 344    6    31     73
#> 345    8    31     85
#> 346   10    31    102
#> 347   12    31    123
#> 348   14    31    138
#> 349   16    31    170
#> 350   18    31    204
#> 351   20    31    235
#> 352   21    31    256
#> 353    0    32     41
#> 354    2    32     49
#> 355    4    32     65
#> 356    6    32     82
#> 357    8    32    107
#> 358   10    32    129
#> 359   12    32    159
#> 360   14    32    179
#> 361   16    32    221
#> 362   18    32    263
#> 363   20    32    291
#> 364   21    32    305
#> 365    0    33     39
#> 366    2    33     50
#> 367    4    33     63
#> 368    6    33     77
#> 369    8    33     96
#> 370   10    33    111
#> 371   12    33    137
#> 372   14    33    144
#> 373   16    33    151
#> 374   18    33    146
#> 375   20    33    156
#> 376   21    33    147
#> 377    0    34     41
#> 378    2    34     49
#> 379    4    34     63
#> 380    6    34     85
#> 381    8    34    107
#> 382   10    34    134
#> 383   12    34    164
#> 384   14    34    186
#> 385   16    34    235
#> 386   18    34    294
#> 387   20    34    327
#> 388   21    34    341
#> 389    0    35     41
#> 390    2    35     53
#> 391    4    35     64
#> 392    6    35     87
#> 393    8    35    123
#> 394   10    35    158
#> 395   12    35    201
#> 396   14    35    238
#> 397   16    35    287
#> 398   18    35    332
#> 399   20    35    361
#> 400   21    35    373
#> 401    0    36     39
#> 402    2    36     48
#> 403    4    36     61
#> 404    6    36     76
#> 405    8    36     98
#> 406   10    36    116
#> 407   12    36    145
#> 408   14    36    166
#> 409   16    36    198
#> 410   18    36    227
#> 411   20    36    225
#> 412   21    36    220
#> 413    0    37     41
#> 414    2    37     48
#> 415    4    37     56
#> 416    6    37     68
#> 417    8    37     80
#> 418   10    37     83
#> 419   12    37    103
#> 420   14    37    112
#> 421   16    37    135
#> 422   18    37    157
#> 423   20    37    169
#> 424   21    37    178
#> 425    0    38     41
#> 426    2    38     49
#> 427    4    38     61
#> 428    6    38     74
#> 429    8    38     98
#> 430   10    38    109
#> 431   12    38    128
#> 432   14    38    154
#> 433   16    38    192
#> 434   18    38    232
#> 435   20    38    280
#> 436   21    38    290
#> 437    0    39     42
#> 438    2    39     50
#> 439    4    39     61
#> 440    6    39     78
#> 441    8    39     89
#> 442   10    39    109
#> 443   12    39    130
#> 444   14    39    146
#> 445   16    39    170
#> 446   18    39    214
#> 447   20    39    250
#> 448   21    39    272
#> 449    0    40     41
#> 450    2    40     55
#> 451    4    40     66
#> 452    6    40     79
#> 453    8    40    101
#> 454   10    40    120
#> 455   12    40    154
#> 456   14    40    182
#> 457   16    40    215
#> 458   18    40    262
#> 459   20    40    295
#> 460   21    40    321
#> 461    0    41     42
#> 462    2    41     51
#> 463    4    41     66
#> 464    6    41     85
#> 465    8    41    103
#> 466   10    41    124
#> 467   12    41    155
#> 468   14    41    153
#> 469   16    41    175
#> 470   18    41    184
#> 471   20    41    199
#> 472   21    41    204
#> 473    0    42     42
#> 474    2    42     49
#> 475    4    42     63
#> 476    6    42     84
#> 477    8    42    103
#> 478   10    42    126
#> 479   12    42    160
#> 480   14    42    174
#> 481   16    42    204
#> 482   18    42    234
#> 483   20    42    269
#> 484   21    42    281
#> 485    0    43     42
#> 486    2    43     55
#> 487    4    43     69
#> 488    6    43     96
#> 489    8    43    131
#> 490   10    43    157
#> 491   12    43    184
#> 492   14    43    188
#> 493   16    43    197
#> 494   18    43    198
#> 495   20    43    199
#> 496   21    43    200
#> 497    0    44     42
#> 498    2    44     51
#> 499    4    44     65
#> 500    6    44     86
#> 501    8    44    103
#> 502   10    44    118
#> 503   12    44    127
#> 504   14    44    138
#> 505   16    44    145
#> 506   18    44    146
#> 507    0    45     41
#> 508    2    45     50
#> 509    4    45     61
#> 510    6    45     78
#> 511    8    45     98
#> 512   10    45    117
#> 513   12    45    135
#> 514   14    45    141
#> 515   16    45    147
#> 516   18    45    174
#> 517   20    45    197
#> 518   21    45    196
#> 519    0    46     40
#> 520    2    46     52
#> 521    4    46     62
#> 522    6    46     82
#> 523    8    46    101
#> 524   10    46    120
#> 525   12    46    144
#> 526   14    46    156
#> 527   16    46    173
#> 528   18    46    210
#> 529   20    46    231
#> 530   21    46    238
#> 531    0    47     41
#> 532    2    47     53
#> 533    4    47     66
#> 534    6    47     79
#> 535    8    47    100
#> 536   10    47    123
#> 537   12    47    148
#> 538   14    47    157
#> 539   16    47    168
#> 540   18    47    185
#> 541   20    47    210
#> 542   21    47    205
#> 543    0    48     39
#> 544    2    48     50
#> 545    4    48     62
#> 546    6    48     80
#> 547    8    48    104
#> 548   10    48    125
#> 549   12    48    154
#> 550   14    48    170
#> 551   16    48    222
#> 552   18    48    261
#> 553   20    48    303
#> 554   21    48    322
#> 555    0    49     40
#> 556    2    49     53
#> 557    4    49     64
#> 558    6    49     85
#> 559    8    49    108
#> 560   10    49    128
#> 561   12    49    152
#> 562   14    49    166
#> 563   16    49    184
#> 564   18    49    203
#> 565   20    49    233
#> 566   21    49    237
#> 567    0    50     41
#> 568    2    50     54
#> 569    4    50     67
#> 570    6    50     84
#> 571    8    50    105
#> 572   10    50    122
#> 573   12    50    155
#> 574   14    50    175
#> 575   16    50    205
#> 576   18    50    234
#> 577   20    50    264
#> 578   21    50    264
#> 
records(db)[[1]]
#>     Chick Diet
#> 1       1    1
#> 13      2    1
#> 25      3    1
#> 37      4    1
#> 49      5    1
#> 61      6    1
#> 73      7    1
#> 85      8    1
#> 96      9    1
#> 108    10    1
#> 120    11    1
#> 132    12    1
#> 144    13    1
#> 156    14    1
#> 168    15    1
#> 176    16    1
#> 183    17    1
#> 195    18    1
#> 197    19    1
#> 209    20    1
#> 221    21    2
#> 233    22    2
#> 245    23    2
#> 257    24    2
#> 269    25    2
#> 281    26    2
#> 293    27    2
#> 305    28    2
#> 317    29    2
#> 329    30    2
#> 341    31    3
#> 353    32    3
#> 365    33    3
#> 377    34    3
#> 389    35    3
#> 401    36    3
#> 413    37    3
#> 425    38    3
#> 437    39    3
#> 449    40    3
#> 461    41    4
#> 473    42    4
#> 485    43    4
#> 497    44    4
#> 507    45    4
#> 519    46    4
#> 531    47    4
#> 543    48    4
#> 555    49    4
#> 567    50    4
records(db)$Chick
#>     Chick Diet
#> 1       1    1
#> 13      2    1
#> 25      3    1
#> 37      4    1
#> 49      5    1
#> 61      6    1
#> 73      7    1
#> 85      8    1
#> 96      9    1
#> 108    10    1
#> 120    11    1
#> 132    12    1
#> 144    13    1
#> 156    14    1
#> 168    15    1
#> 176    16    1
#> 183    17    1
#> 195    18    1
#> 197    19    1
#> 209    20    1
#> 221    21    2
#> 233    22    2
#> 245    23    2
#> 257    24    2
#> 269    25    2
#> 281    26    2
#> 293    27    2
#> 305    28    2
#> 317    29    2
#> 329    30    2
#> 341    31    3
#> 353    32    3
#> 365    33    3
#> 377    34    3
#> 389    35    3
#> 401    36    3
#> 413    37    3
#> 425    38    3
#> 437    39    3
#> 449    40    3
#> 461    41    4
#> 473    42    4
#> 485    43    4
#> 497    44    4
#> 507    45    4
#> 519    46    4
#> 531    47    4
#> 543    48    4
#> 555    49    4
#> 567    50    4

# subsetting first isn't recommended: removes foreign key
# reference as mentions, and you need to subset again anyway
records(db[[1]])[[1]]
#>     Chick Diet
#> 1       1    1
#> 13      2    1
#> 25      3    1
#> 37      4    1
#> 49      5    1
#> 61      6    1
#> 73      7    1
#> 85      8    1
#> 96      9    1
#> 108    10    1
#> 120    11    1
#> 132    12    1
#> 144    13    1
#> 156    14    1
#> 168    15    1
#> 176    16    1
#> 183    17    1
#> 195    18    1
#> 197    19    1
#> 209    20    1
#> 221    21    2
#> 233    22    2
#> 245    23    2
#> 257    24    2
#> 269    25    2
#> 281    26    2
#> 293    27    2
#> 305    28    2
#> 317    29    2
#> 329    30    2
#> 341    31    3
#> 353    32    3
#> 365    33    3
#> 377    34    3
#> 389    35    3
#> 401    36    3
#> 413    37    3
#> 425    38    3
#> 437    39    3
#> 449    40    3
#> 461    41    4
#> 473    42    4
#> 485    43    4
#> 497    44    4
#> 507    45    4
#> 519    46    4
#> 531    47    4
#> 543    48    4
#> 555    49    4
#> 567    50    4